added newstories and topstories, snake game, optimized Dockerfile, simpler main loop

+++ scripts/main_loop.sh
@@ -1,27 +1,34 @@
 #! /bin/bash
 
-
 REFRESH_INTERVAL=${REFRESH_INTERVAL:-3600}
+COUNTER=0
+RESET_COUNTER=$((REFRESH_INTERVAL*24))
 
 nginx
-echo Generating index.html .. > index.html
+echo Refresh interval: ${REFRESH_INTERVAL}s
+echo Cache Refresh interval: ${RESET_COUNTER}s
+echo Please enjoy a little Snake game while generating index.html
+
+rm -rf /home/user/cache/hn
+mkdir -p /home/user/cache/hn
 
 while true
 do
     echo "-------------------------------------------"
-    echo "Start: $(date)"
-    node --trace-warnings --experimental-modules ./bin/generate-html.mjs --hacker-news=hn 2>&1
+    echo "Regenerating index.html: $(date)"
+    node --no-warnings ./bin/generate-html.mjs --hacker-news=hn --feed=newstories --target=index.html 2>&1
+    node --no-warnings ./bin/generate-html.mjs --hacker-news=hn --feed=topstories --target=top.html 2>&1
     echo "Done: $(date)"
-    sleep $REFRESH_INTERVAL
 
-    mkdir -p /home/user/cache/hn2
-    rm -rf /home/user/cache/hn2
-    echo "-------------------------------------------"
-    echo "Start: $(date)"
+    COUNTER=$((COUNTER + REFRESH_INTERVAL))
 
-    node --trace-warnings --experimental-modules ./bin/generate-html.mjs --hacker-news=hn2 2>&1
-    rm -rf /home/user/cache/hn
-    echo "Done: $(date)"
+    if [ $COUNTER -ge $RESET_COUNTER ] 
+    then
+        echo "Deleting cache!"
+        rm -rf /home/user/cache/hn
+        mkdir -p /home/user/cache/hn      
+        COUNTER=0
+    fi
+
     sleep $REFRESH_INTERVAL
 done
-
